+++ apps/decodex/src/recovery/tests/stale_active/telemetry.rs
@@ -1,0 +1,398 @@
+use super::*;
+
+#[test]
+fn stale_active_diagnose_allows_dead_orphan_thread_runtime_telemetry() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let queue_label = tracker::automation_queue_label("pubfi");
+	let mut issue = sample_issue_with_labels("In Progress", &[active_label, queue_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
+	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
+	assert!(diagnostic.evidence.contains(&String::from("process_not_alive")));
+	assert!(diagnostic.evidence.contains(&String::from("stale_active_control_channel_present")));
+	assert!(
+		diagnostic
+			.evidence
+			.contains(&String::from("only_stale_active_or_failed_control_evidence_present"))
+	);
+}
+
+#[test]
+fn stale_active_diagnose_allows_app_server_no_progress_failure_evidence() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let queue_label = tracker::automation_queue_label("pubfi");
+	let mut issue = sample_issue_with_labels("In Progress", &[active_label, queue_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
+	append_app_server_no_progress_failure_evidence(&store, &issue.id);
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
+	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
+	assert!(
+		diagnostic
+			.evidence
+			.contains(&String::from("only_stale_active_or_failed_control_evidence_present"))
+	);
+}
+
+#[test]
+fn stale_active_diagnose_blocks_harness_outcome_with_pr_progress() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let queue_label = tracker::automation_queue_label("pubfi");
+	let mut issue = sample_issue_with_labels("In Progress", &[active_label, queue_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
+	append_harness_outcome_with_pr_progress(&store, &issue.id);
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
+	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
+	assert!(!diagnostic.recoverable());
+}
+
+#[test]
+fn stale_active_diagnose_blocks_harness_outcome_with_review_progress() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let queue_label = tracker::automation_queue_label("pubfi");
+	let mut issue = sample_issue_with_labels("In Progress", &[active_label, queue_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
+	append_harness_outcome_with_review_progress(&store, &issue.id);
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
+	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
+	assert!(!diagnostic.recoverable());
+}
+
+#[test]
+fn stale_active_diagnose_blocks_harness_outcome_with_validation_progress() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let queue_label = tracker::automation_queue_label("pubfi");
+	let mut issue = sample_issue_with_labels("In Progress", &[active_label, queue_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
+	append_harness_outcome_with_validation_progress(&store, &issue.id);
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
+	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
+	assert!(!diagnostic.recoverable());
+}
+
+#[test]
+fn stale_active_diagnose_blocks_no_diff_guardrail_with_delta() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let queue_label = tracker::automation_queue_label("pubfi");
+	let mut issue = sample_issue_with_labels("In Progress", &[active_label, queue_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
+	append_no_diff_guardrail_event(&store, &issue.id, true, false);
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
+	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
+	assert!(!diagnostic.recoverable());
+}
+
+#[test]
+fn stale_active_diagnose_allows_app_server_phase_goal_recovery_telemetry() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let queue_label = tracker::automation_queue_label("pubfi");
+	let mut issue = sample_issue_with_labels("In Progress", &[active_label, queue_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
+	append_phase_goal_recovery_event(
+		&store,
+		&issue.id,
+		"implement_to_validation_ready",
+		"app_server_dynamic_tool_protocol_failure",
+	);
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
+	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
+}
+
+#[test]
+fn stale_active_diagnose_blocks_repo_gate_phase_goal_recovery_telemetry() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let queue_label = tracker::automation_queue_label("pubfi");
+	let mut issue = sample_issue_with_labels("In Progress", &[active_label, queue_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
+	append_phase_goal_recovery_event(
+		&store,
+		&issue.id,
+		"implement_to_validation_ready",
+		"repo_gate_verify_failed",
+	);
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
+	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
+	assert!(!diagnostic.recoverable());
+}
+
+#[test]
+fn stale_active_diagnose_blocks_repair_phase_goal_recovery_telemetry() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let queue_label = tracker::automation_queue_label("pubfi");
+	let mut issue = sample_issue_with_labels("In Progress", &[active_label, queue_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
+	append_phase_goal_recovery_event(
+		&store,
+		&issue.id,
+		"repair_accepted_review_findings",
+		"app_server_dynamic_tool_protocol_failure",
+	);
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
+	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
+	assert!(!diagnostic.recoverable());
+}
+
+#[test]
+fn stale_active_diagnose_blocks_clean_worktree_with_unmerged_commits() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let mut issue = sample_issue_with_labels("Todo", &[active_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	init_git_repo(&worktree_path);
+	run_git(&worktree_path, &["checkout", "-B", "main"]);
+	commit_test_file(&worktree_path, "README.md", "base\n", "base");
+	run_git(&worktree_path, &["checkout", "-b", "x/pubfi-pub-1626"]);
+	commit_test_file(&worktree_path, "source.rs", "fn retained_progress() {}\n", "progress");
+	store
+		.record_run_attempt("run-1626", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-1626",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree mapping should record");
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.worktree_state, "unmerged_commits_present");
+	assert!(diagnostic.blockers.contains(&String::from("worktree_unmerged_commits_present")));
+	assert!(!diagnostic.recoverable());
+}
+
+#[test]
+fn stale_active_diagnose_blocks_clean_git_worktree_without_default_branch() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let store = StateStore::open_in_memory().expect("state store should open");
+	let workflow = sample_workflow();
+	let active_label = tracker::automation_active_label("pubfi");
+	let mut issue = sample_issue_with_labels("Todo", &[active_label]);
+	let worktree_path = temp_dir.path().join("PUB-1626");
+
+	issue.identifier = String::from("PUB-1626");
+	init_git_repo(&worktree_path);
+	store
+		.record_run_attempt("run-1626", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-1626",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree mapping should record");
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
+	let diagnostics = super::super::super::diagnose_stale_active_issues(
+		"pubfi",
+		&workflow,
+		temp_dir.path(),
+		&store,
+		&tracker,
+		Some("PUB-1626"),
+		super::super::super::RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.first().expect("diagnostic should exist");
+
+	assert_eq!(diagnostic.worktree_state, "default_branch_unavailable");
+	assert!(diagnostic.blockers.contains(&String::from("worktree_default_branch_unavailable")));
+	assert!(!diagnostic.recoverable());
+}
